% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ComputeEquilibriumCodonFrequencies}
\alias{ComputeEquilibriumCodonFrequencies}
\title{Function to plot frequency of distribution of different Wi given selac parameters}
\usage{
ComputeEquilibriumCodonFrequencies(nuc.model = "JC", base.freqs = rep(0.25,
  4), nsites = 1, C = 4, Phi = 0.5, q = 4e-07, Ne = 5e+06,
  alpha = 1.83, beta = 0.1, gamma = 0.0003990333,
  include.stop.codon = TRUE, numcode = 1, diploid = TRUE,
  flee.stop.codon.rate = 0.9999999)
}
\description{
Function to plot frequency of distribution of different Wi given selac parameters
}

