% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selacSim.R
\name{FullWrightFisherSimulator}
\alias{FullWrightFisherSimulator}
\title{Simulate using an individual based model}
\usage{
FullWrightFisherSimulator(phy, root.states = NULL, phi.mean, phi.sd = 0,
  Ne = 100, GTR.rates, recombination.rate, optimal.aa.switch.rate,
  diploid = TRUE, aa.properties = NULL, population = NULL,
  brlen.conversion = 1e+06, burnin.generations = 1000)
}
\arguments{
\item{phy}{A phylogram}

\item{root.states}{The starting sequence}

\item{phi.mean}{The mean value of phi across sites}

\item{phi.sd}{The standard deviation of phi across sites}

\item{Ne}{Effective population size. Must be an integer.}

\item{recombination.rate}{The rate at which there is recombination}

\item{optimal.aa.switch.rate}{The rate at which the optimal amino acid changes}

\item{aa.properties}{The aa.properties}

\item{population}{Rather than a single sequence to start the sim, allows use of a population}

\item{brlen.conversion}{How many generations correspond to one unit of branch length on a tree}

\item{burnin.generations}{How many generations to simulate on before actually starting the sim.}

\item{GTR.rates}{rates in the usual order for selac}

\item{diploid}{If TRUE, simulate using a diploid population}
}
\description{
Simulate using an individual based model
}

