% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalRecon.R
\name{GetMarginalGene}
\alias{GetMarginalGene}
\title{Get marginal reconstruction a single}
\usage{
GetMarginalGene(pars, codon.data, phy, codon.freq.by.aa = NULL,
  codon.freq.by.gene = NULL, aa.optim_array, numcode, nuc.model = "UNREST",
  include.gamma = FALSE, gamma.type = "quadrature", ncats = 4,
  k.levels = 0, diploid = TRUE, taxon.to.drop = NULL,
  aa.properties = NULL, data.type = "codon")
}
\arguments{
\item{pars}{A vector of parameters used for the simulation. They are ordered as follows: C.q.phi, alpha, beta, Ne, base.freqs for A C G, and the rates for the nucleotide model.}

\item{phy}{The phylogenetic tree with branch lengths.}

\item{codon.freq.by.aa}{A matrix of codon frequencies for each possible optimal amino acid. Rows are aa (including stop codon), cols are codons.}

\item{codon.freq.by.gene}{A matrix of codon frequencies for each gene.}

\item{aa.optim_array}{A vector of optimal amino acids for each site to be simulated.}

\item{numcode}{The ncbi genetic code number for translation. By default the standard (numcode=1) genetic code is used.}

\item{nuc.model}{Indicates what type nucleotide model to use. There are three options: "JC", "GTR", or "UNREST".}

\item{include.gamma}{A logical indicating whether or not to include a discrete gamma model.}

\item{gamma.type}{Indicates what type of gamma distribution to use. Options are "quadrature" after the Laguerre quadrature approach of Felsenstein 2001 or median approach of Yang 1994.}

\item{ncats}{The number of discrete categories.}

\item{k.levels}{Provides how many levels in the polynomial. By default we assume a single level (i.e., linear).}

\item{diploid}{A logical indicating whether or not the organism is diploid or not.}

\item{taxon.to.drop}{A single taxon (defined by number in phy object) to be removed from the reconstruction.}

\item{aa.properties}{User-supplied amino acid distance properties. By default we assume Grantham (1974) properties.}

\item{data.type}{The data type being tested. Options are "codon" or "nucleotide".}
}
\description{
Calculates the marginal probability of each codon at all sites across a single gene
}
\details{
Provides marginal probabilities for all nodes across a single gene.
}

