\name{internal}
\alias{CreateAADistanceMatrix}
\alias{CreateNucleotideMutationMatrix}
\alias{CreateCodonMutationMatrix}
\alias{CompareVectors}
\alias{GetPairwiseProteinFixationProbabilityArbitraryLength}
\alias{GetPairwiseProteinFixationProbabilitySingleSite}
\alias{TranslateCodon}
\alias{GetProteinProteinDistance}
\alias{CreateCodonFixationRateMatrix}
\alias{CreateAAFixationMatrix}
\alias{GetLikelihoodSAC_AAForSingleCharGivenOptimum}
\docType{data}
\title{
Internal functions
}
\description{
These are functions used internally which will generally not be called by the user.
}

\keyword{internal}
