% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{LineMutationFitnessSpectra}
\alias{LineMutationFitnessSpectra}
\title{Generates a line for mutation fitness spectra}
\usage{
LineMutationFitnessSpectra(mutation.fitness.object, optimal.aa = NULL)
}
\description{
Generates a line for mutation fitness spectra
}

