% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotGeneSiteInfo}
\alias{PlotGeneSiteInfo}
\title{Function to plot info by site in a gene}
\usage{
PlotGeneSiteInfo(info.by.site, aa.properties = NULL)
}
\arguments{
\item{info.by.site}{The output of GetGeneSiteInfo}
}
\description{
Function to plot info by site in a gene
}

