% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotMutationFitnessSpectra}
\alias{PlotMutationFitnessSpectra}
\title{Plot fitness of mutations, weighted by frequency of those mutations}
\usage{
PlotMutationFitnessSpectra(mutation.fitness.object.list, values,
  optimal.aa = NULL, palette = "Set1", lwd = 2, ...)
}
\arguments{
\item{mutation.fitness.object.list}{List that contains multiple objects from ComputeMutationFitnesses() calls}

\item{values}{The vector of labels for each matrix (i.e., different Phi values)}

\item{optimal.aa}{Single letter code for the optimal aa. If NULL, integrates across aa.}

\item{palette}{Color palette to use from RColorBrewer}

\item{lwd}{Line width}

\item{...}{other arguments to pass to plot()}
}
\description{
Plot fitness of mutations, weighted by frequency of those mutations
}
