% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotPerAAFitness}
\alias{PlotPerAAFitness}
\title{Function to plot a distribution of fitnesses across amino acids}
\usage{
PlotPerAAFitness(aa.fitness.matrices, values, optimal.aa = NULL,
  palette = "Set1", lwd = 2, include.stop.codon = FALSE,
  type = "histogram", fitness = TRUE, ...)
}
\arguments{
\item{aa.fitness.matrices, }{A 3d array of aa.fitness.matrix returned from ComputeEquilibriumAAFitness (first element in return)}

\item{values}{The vector of labels for each matrix (i.e., different Phi values)}

\item{optimal.aa}{Single letter code for the optimal aa. If NULL, integrates across aa.}

\item{palette}{Color palette to use from RColorBrewer}

\item{lwd}{Line width}

\item{include.stop.codon}{Include stop codons}

\item{type}{If "histogram", do a histogram plot; if "density", do a density plot}

\item{fitness}{If TRUE, plot W; if FALSE, plot S (= 1 - W)}

\item{...}{Other paramters to pass to plot()}
}
\description{
Function to plot a distribution of fitnesses across amino acids
}
\examples{
phi.vector <- c(0.0000000001, 0.01, .1, 0.5, 2)
aa.fitness.matrices <- array(dim=c(21, 20, length(phi.vector)))
for (i in sequence(length(phi.vector))) {
 local.matrix <- ComputeEquilibriumAAFitness(Phi=phi.vector[i])$aa.fitness.matrix
 aa.fitness.matrices[,,i] <- local.matrix
 dimnames(aa.fitness.matrices) <- list(rownames(local.matrix), colnames(local.matrix), NULL)
}
values = paste("Phi = ", phi.vector, sep="")
PlotPerAAFitness(aa.fitness.matrices, values, optimal.aa="L")
}

