% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selac.R
\name{SelacLargeOptimize}
\alias{SelacLargeOptimize}
\title{Efficient optimization of the SELAC model}
\usage{
SelacLargeOptimize(codon.data.path, n.partitions = NULL, phy,
  data.type = "codon", edge.length = "optimize", edge.linked = TRUE,
  optimal.aa = "optimize", nuc.model = "GTR", include.gamma = FALSE,
  ncats = 4, numcode = 1, diploid = TRUE, k.levels = 0,
  aa.properties = NULL, verbose = FALSE, n.cores = NULL,
  max.tol = .Machine$double.eps^0.5, max.evals = 1e+06, max.restarts = 3,
  user.optimal.aa = NULL, fasta.rows.to.keep = NULL,
  recalculate.starting.brlen = TRUE, output.by.restart = TRUE,
  output.restart.filename = "restartResult")
}
\arguments{
\item{codon.data.path}{Provides the path to the directory containing the gene specific fasta files of coding data. Must have a ".fasta" line ending.}

\item{n.partitions}{The number of partitions to analyze. The order is based on the Unix order of the fasta files in the directory.}

\item{phy}{The phylogenetic tree to optimize the model parameters.}

\item{data.type}{The data type being tested. Options are "codon" or "nucleotide".}

\item{edge.length}{Indicates whether or not edge lengths should be optimized. By default it is set to "optimize", other option is "fixed", which user-supplied branch lengths.}

\item{edge.linked}{A logical indicating whether or not edge lengths should be optimized separately for each gene. By default, a single set of each lengths is optimized for all genes.}

\item{optimal.aa}{Indicates what type of optimal.aa should be used. There are four options: "none", "majrule", "optimize", or "user".}

\item{nuc.model}{Indicates what type nucleotide model to use. There are three options: "JC", "GTR", or "UNREST".}

\item{include.gamma}{A logical indicating whether or not to include a discrete gamma model.}

\item{ncats}{The number of discrete categories.}

\item{numcode}{The The ncbi genetic code number for translation. By default the standard (numcode=1) genetic code is used.}

\item{diploid}{A logical indicating whether or not the organism is diploid or not.}

\item{k.levels}{Provides how many levels in the polynomial. By default we assume a single level (i.e., linear).}

\item{aa.properties}{User-supplied amino acid distance properties. By default we assume Grantham (1974) properties.}

\item{verbose}{Logical indicating whether each iteration be printed to the screen.}

\item{n.cores}{The number of cores to run the analyses over.}

\item{max.tol}{Supplies the relative optimization tolerance.}

\item{max.evals}{Supplies the max number of iterations tried during optimization.}

\item{max.restarts}{Supplies the number of random restarts.}

\item{user.optimal.aa}{If optimal.aa is set to "user", this option allows for the user-input optimal amino acids. Must be a list. To get the proper order of the partitions see "GetPartitionOrder" documentation.}

\item{fasta.rows.to.keep}{Indicates which rows to remove in the input fasta files.}

\item{recalculate.starting.brlen}{Whether to use given branch lengths in the starting tree or recalculate them.}

\item{output.by.restart}{Logical indicating whether or not each restart is saved to a file. Default is TRUE.}

\item{output.restart.filename}{Designates the file name for each random restart.}
}
\description{
Efficient optimization of model parameters under the SELAC model
}
\details{
This is the exact same implementation as in SelacOptimize(), except here we optimize parameters across each gene separately while keeping the shared parameters, alpha, beta, edge lengths, constant across genes. We then optimize alpha, beta, and the edge lengths while keeping parameters for each gene fixed. This approach is potentially more efficient than simply optimizing all parameters simultaneously, especially if fitting models across 100's of genes.
}

