% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selon.R
\name{SelonLargeOptimize}
\alias{SelonLargeOptimize}
\title{Efficient optimization of the SELON model}
\usage{
SelonLargeOptimize(nuc.data.path, n.partitions = NULL, phy,
  edge.length = "optimize", edge.linked = TRUE, optimal.nuc = "majrule",
  nuc.model = "GTR", diploid = TRUE, verbose = FALSE, n.cores = NULL,
  max.tol = .Machine$double.eps^0.5, max.evals = 1e+06, max.restarts = 10,
  output.by.restart = TRUE, output.restart.filename = "restartResult",
  fasta.rows.to.keep = NULL)
}
\arguments{
\item{nuc.data.path}{Provides the path to the directory containing the gene specific fasta files that contains the nucleotide data.}

\item{n.partitions}{The number of partitions to analyze. The order is based on the Unix order of the fasta files in the directory.}

\item{phy}{The phylogenetic tree to optimize the model parameters.}

\item{edge.length}{Indicates whether or not edge lengths should be optimized. By default it is set to "optimize", other option is "fixed", which user-supplied branch lengths.}

\item{edge.linked}{A logical indicating whether or not edge lengths should be optimized separately for each gene. By default, a single set of each lengths is optimized for all genes.}

\item{optimal.nuc}{Indicates what type of optimal.nuc should be used. At the moment there is only a single option: "majrule".}

\item{nuc.model}{Indicates what type nucleotide model to use. There are three options: "JC", "GTR", or "UNREST".}

\item{diploid}{A logical indicating whether or not the organism is diploid or not.}

\item{verbose}{Logical indicating whether each iteration be printed to the screen.}

\item{n.cores}{The number of cores to run the analyses over.}

\item{max.tol}{Supplies the relative optimization tolerance.}

\item{max.evals}{Supplies the max number of iterations tried during optimization.}

\item{max.restarts}{Supplies the number of random restarts.}

\item{output.by.restart}{Logical indicating whether or not each restart is saved to a file. Default is TRUE.}

\item{output.restart.filename}{Designates the file name for each random restart.}

\item{fasta.rows.to.keep}{Indicates which rows to remove in the input fasta files.}
}
\description{
Efficient optimization of model parameters under the SELON model
}
\details{
This is the exact same implementation as in SelonOptimize(), except here we optimize parameters across each gene separately while keeping the shared parameters, the edge lengths, constant across genes. We then optimize edge lengths while keeping parameters for each gene fixed. This approach is potentially more efficient than simply optimizing all parameters simultaneously, especially if fitting models across 100's of genes.
}

