% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportRegion.R
\name{SupportRegion}
\alias{SupportRegion}
\title{Adaptive sampling of support region}
\usage{
SupportRegion(selac.obj, n.points = 10000, scale.int = 0.1,
  desired.delta = 2, parallel.type = "by.gene", n.cores = NULL,
  verbose = TRUE, fasta.rows.to.keep = NULL)
}
\arguments{
\item{selac.obj}{An object of class "selac" that contains the MLE from a given model run.}

\item{n.points}{Indicates the number of points to sample.}

\item{scale.int}{The scaling multiplier that defines the interval to randomly sample. By default it is set to 0.1, meaning that values are drawn at random along an interval that encompasses 10 percent above and below the MLE.}

\item{desired.delta}{Defines the number of lnL units away from the MLE to include. By default the value is set to 2.}

\item{parallel.type}{Designates whether a parallel run should occur by gene ("by.gene") or by site ("by.site").}

\item{n.cores}{The number of cores to run analyses over.}

\item{verbose}{A logical indicating whether progress should be printed to the screen. The default is TRUE.}

\item{fasta.rows.to.keep}{Indicates which rows to remove in the input fasta files.}
}
\description{
Adaptively samples points for each parameter to obtain an of the confidence intervals
}
\details{
This function provides a means for sampling the likelihood surface quickly to estimate confidence intervals that reflect the uncertainty in the MLE. The function starts with the MLE from the hisse run. It then uses a scaling multiplier to generate an interval by which to randomly alter each parameter. However, the algorithm was designed to \dQuote{feel} the boundaries of the random search. In other words, when the algorithm begins to sample the hinterlands of the surface, it will know to restrict the boundary to allow sampling of more reasonable values based on the currently sampled set. The goal of this sampling process is to find points within some desired distance from the MLE; by default we assume this distance is 2 lnLik. The confidence interval can be estimated directly from these points. The full set of points tried are also provided and can be used to generate contour plots (though, it is not entirely straightforward to do so -- but certainly doable).
}

