% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotEquilbriumCodonDistribution}
\alias{PlotEquilbriumCodonDistribution}
\title{Function to plot a distribution of frequencies of codons given a 3d array of equilibrium frequency matrices}
\usage{
PlotEquilbriumCodonDistribution(eq.freq.matrices, values,
  palette = "Set1", lwd = 2, ...)
}
\arguments{
\item{eq.freq.matrices}{A 3d array of eq.freq.matrix returned from ComputeEquilibriumFrequencies}

\item{values}{The vector of labels for each matrix (i.e., different Phi values)}

\item{palette}{Color palette to use from RColorBrewer}

\item{lwd}{Line width}

\item{...}{Other paramters to pass to plot()}
}
\description{
Function to plot a distribution of frequencies of codons given a 3d array of equilibrium frequency matrices
}
