% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotGeneSiteInfo}
\alias{PlotGeneSiteInfo}
\title{Function to plot info by site in a gene}
\usage{
PlotGeneSiteInfo(all.info, aa.properties = NULL, mean.width = 10)
}
\arguments{
\item{all.info}{The output of GetGeneSiteInfo}

\item{aa.properties}{The aa.properties you want to use; if NULL, uses Grantham}

\item{mean.width}{Sliding window width}
}
\description{
Function to plot info by site in a gene
}
